Refactor ISO class into Xorriso
Summary:
  * Move the ISO class into the Xorriso class
    to make it more flexible.
